BAU: Stop creating sns topic in every integration test
 Sns topics are only used in a single place now in auth's code, but the sns topic extension was included in the base integration test, it was being set up before every single test. We saw flakiness on the setup of the tests, so only using it where it is needed should cut down on the flakiness
