+++ .secrets.baseline
@@ -217,7 +217,7 @@
         "filename": "account-management-integration-tests/src/test/java/uk/gov/di/accountmanagement/api/BulkRemoveAccountHandlerIntegrationTest.java",
         "hashed_secret": "cbfdac6008f9cab4083784cbd1874f76618d2a97",
         "is_verified": false,
-        "line_number": 36,
+        "line_number": 39,
         "is_secret": false
       }
     ],
@@ -1610,7 +1610,7 @@
         "filename": "shared-test/src/main/java/uk/gov/di/authentication/sharedtest/basetest/HandlerIntegrationTest.java",
         "hashed_secret": "46c0222178bd5e9795849bf479f7adfc236ff9dc",
         "is_verified": false,
-        "line_number": 54,
+        "line_number": 52,
         "is_secret": false
       }
     ],
